pyproject.toml
[tool.poetry]
name = "plants-pro-api"
version = "0.1.0"
description = ""
authors = ["jumashafara <jumashafara0@gmail.com>"]
readme = "README.md"
packages = [{include = "plants_pro_api"}]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.5"
dj-database-url = "^1.2.0"
psycopg2-binary = "^2.9.5"
whitenoise = {extras = ["brotli"], version = "^6.3.0"}
gunicorn = "^20.1.0"
django-rest-framework = "^0.1.0"
opencv-python = "^4.7.0.68"
pathlib = "^1.0.1"
django-cors-headers = "^3.13.0"
numpy = "^1.24.2"
pillow = "^9.4.0"
tensorflow-hub = "^0.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
